+++ lib/rough_notation.dart
@@ -9,6 +9,6 @@ export 'src/annotations/rough_circle_annotation.dart';
 export 'src/annotations/rough_highlight_annotation.dart';
 export 'src/annotations/rough_bracket_annotation.dart';
 
-// all controllers
+/// all controllers
 export 'src/controllers/rough_annotation_registry.dart';
 export 'src/controllers/rough_annotation_controller.dart';