+++ playground/lib/widgets/custom_button.dart
@@ -21,9 +21,9 @@ class CustomButton extends StatelessWidget {
         ),
         child: Text(
           text?.toUpperCase() ?? 'ANNOTATE',
-          style: ShadTheme.of(context).textTheme.p.copyWith(
-            fontWeight: FontWeight.w500
-          ),
+          style: ShadTheme.of(
+            context,
+          ).textTheme.p.copyWith(fontWeight: FontWeight.w500),
         ),
       ),
     );